refactor: use AnimatePulsingDot island for Liveblogs
